Update setup.py
This is a python file used by the cx_freeze module. It takes the file you want to convert, the name you want your exe to have, and any packages or files your program needs. 
Use by opening a terminal in your directory and typing: python setup.py build (for an exe), or: pyhton setup.py bdist_msi (for a windows installer)